+++ experimental/transport-otlp-http/src/payload/transform/toErrorLogRecord.test.ts
@@ -77,7 +77,7 @@ const matchErrorLogRecord = {
       },
     },
     {
-      key: 'http.url',
+      key: 'url.full',
       value: {
         stringValue: 'http://localhost:5173',
       },
@@ -135,19 +135,19 @@ const matchErrorLogRecord = {
       },
     },
     {
-      key: 'enduser.id',
+      key: 'user.id',
       value: { stringValue: 'user-abc123' },
     },
     {
-      key: 'enduser.name',
+      key: 'user.name',
       value: { stringValue: 'user-joe' },
     },
     {
-      key: 'enduser.email',
+      key: 'user.email',
       value: { stringValue: 'user@example.com' },
     },
     {
-      key: 'enduser.attributes',
+      key: 'user.attributes',
       value: {
         kvlistValue: {
           values: [