+++ experimental/transport-otlp-http/src/payload/transform/toEventLogRecord.test.ts
@@ -60,7 +60,7 @@ const matchEventLogRecord = {
       },
     },
     {
-      key: 'http.url',
+      key: 'url.full',
       value: {
         stringValue: 'http://localhost:5173',
       },
@@ -118,19 +118,19 @@ const matchEventLogRecord = {
       },
     },
     {
-      key: 'enduser.id',
+      key: 'user.id',
       value: { stringValue: 'user-abc123' },
     },
     {
-      key: 'enduser.name',
+      key: 'user.name',
       value: { stringValue: 'user-joe' },
     },
     {
-      key: 'enduser.email',
+      key: 'user.email',
       value: { stringValue: 'user@example.com' },
     },
     {
-      key: 'enduser.attributes',
+      key: 'user.attributes',
       value: {
         kvlistValue: {
           values: [