+++ packages/web-tracing/src/sessionSpanProcessor.ts
@@ -1,9 +1,11 @@
 import type { Context } from '@opentelemetry/api';
 import type { ReadableSpan, Span, SpanProcessor } from '@opentelemetry/sdk-trace-web';
+// False positive. Package can be resolved.
+// eslint-disable-next-line import/no-unresolved
+import { ATTR_SESSION_ID } from '@opentelemetry/semantic-conventions/incubating';
 
 import type { Metas } from '@grafana/faro-web-sdk';
 
-// adds Faro session id to every span
 export class FaroSessionSpanProcessor implements SpanProcessor {
   constructor(
     private processor: SpanProcessor,
@@ -18,6 +20,10 @@ export class FaroSessionSpanProcessor implements SpanProcessor {
     const session = this.metas.value.session;
 
     if (session?.id) {
+      span.attributes[ATTR_SESSION_ID] = session.id;
+      /**
+       * @deprecated will be removed in the future and has been replaced by ATTR_SESSION_ID (session.id)
+       */
       span.attributes['session_id'] = session.id;
     }