+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## Next
 
+- Dependencies (`@grafana/faro-web-tracing`): upgrade otel deps (#670).
+  - Note: some attributes have been changed due to otel semantic attributes spec or are now aligned
+    with it. For the web-tracing package we provide both attribute versions for now:
+    - `deployment.environment` is now deprecated and will be replaced by
+      `deployment.environment.name`.
+    - `session_id` is now deprecated and will be replaced by `session.id`
+- Dependencies (`@grafana/faro-core`): upgrade otel deps (#670).
+- Dependencies (`@grafana/faro-transport-otlp-http [experimental]`): upgrade otel deps (#670).
+  - Note: some attributes have been changed due to otel semantic attributes spec:
+    - `enduser.id` is replaced by `user.id`
+    - `enduser.name` is replaced by `user.username`,
+    - `enduser.email` is replaced by `user.email`,
+    - `enduser.attributes` is replaced by `user.attributes`,
+    - `http.url` is replaced by `url.full`
+    - `deployment.environment` is replaced by `deployment.environment.name`
+
 ## 1.9.1
 
 - Fix (`@grafana/faro-transport-otlp-http [experimental]`): Properly consume response body (#664).