+++ experimental/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Next
 
+- Dependencies (`@grafana/faro-transport-otlp-http`): upgrade otel deps (#670).
+  - Note: some attributes have been changed due to otel semantic attributes spec:
+    - `enduser.id` is replaced by `user.id`
+    - `http.url` is replaced by `url.full`
+    - `deployment.environment` is replaced by `deployment.environment.name`
+
 ## 1.9.1
 
 - Fix (`@grafana/faro-transport-otlp-http`): Properly consume response body (#664).